+++ packages/web-components/@sfx/autocomplete/stories/index.ts
@@ -1,8 +1,17 @@
 import { storiesOf } from '@storybook/html';
 import { withKnobs, text } from '@storybook/addon-knobs';
-import { getDisplayCode, autocompleteReceivedResultsEvent, autocompleteResults, hidePrompt } from '../../../../../.storybook/common';
+import {
+  AUTOCOMPLETE_RESPONSE,
+  AutocompleteResultGroup,
+  AutocompleteSearchTermItem,
+} from '@sfx/events';
+import {
+  getDisplayCode,
+  autocompleteResponseEvent,
+  autocompleteResults,
+  hidePrompt,
+} from '../../../../../.storybook/common';
 import '../src/index';
-import { AUTOCOMPLETE_RECEIVED_RESULTS_EVENT } from '../src/index';
 
 const autocompleteNotesIntro = `
 # SF-X Autocomplete Component
@@ -15,8 +24,7 @@ const autocompleteNotesIntro = `
 
 ## Demonstrated in this story`
 
-// TODO: add results type to function when imported from @sfx/events repo
-function getAutocompleteComponent(results = []): string {
+function getAutocompleteComponent(results: AutocompleteResultGroup<AutocompleteSearchTermItem>[] = []): string {
   const optionalTitle = text('Optional Title', 'Autocomplete Results');
 
   if (results.length > 0) {
@@ -76,7 +84,7 @@ storiesOf('Components|Autocomplete', module)
   .add(
     'Rendering with event payload',
     () => {
-      hidePrompt(AUTOCOMPLETE_RECEIVED_RESULTS_EVENT);
+      hidePrompt(AUTOCOMPLETE_RESPONSE);
       const autocomplete = getAutocompleteComponent();
 
       return `
@@ -86,14 +94,14 @@ storiesOf('Components|Autocomplete', module)
      `;
     },
     {
-      customEvents: [autocompleteReceivedResultsEvent],
+      customEvents: [autocompleteResponseEvent],
       notes: {
         markdown: `
           ${autocompleteNotesIntro}
 
-            ### The SF-X Autocomplete component updates with autocomplete data in response to the \`${AUTOCOMPLETE_RECEIVED_RESULTS_EVENT}\` event.**
+            ### The SF-X Autocomplete component updates with autocomplete data in response to the \`${AUTOCOMPLETE_RESPONSE}\` event.**
             * To emit the event in this story:
-              1. Visit the **Custom Events** tab and locate the \`${AUTOCOMPLETE_RECEIVED_RESULTS_EVENT}\` event.
+              1. Visit the **Custom Events** tab and locate the \`${AUTOCOMPLETE_RESPONSE}\` event.
               2. Click "emit".
               3. Observe that the component is updated with the payload of the event.