+++ packages/web-components/@sfx/product/src/index.ts
@@ -1,2 +1,2 @@
-export { default as Product, ProductModel, ProductVariantModel, ProductVariantsModel } from './product';
+export { default as Product } from './product';
 export { default as Variant } from './variant';