+++ packages/web-components/@sfx/product/stories/index.ts
@@ -1,11 +1,10 @@
 import { storiesOf } from '@storybook/html';
 import { withKnobs, text } from '@storybook/addon-knobs';
-import { ProductModel } from '@sfx/product';
-
+import { Product } from '@sfx/events';
 import { getDisplayCode, getProducts } from '../../../../../.storybook/common';
 import '../src/index';
 
-function getProductComponent(product: ProductModel = {}, codeSnippet: boolean = false): string {
+function getProductComponent(product: Product = {}, codeSnippet: boolean = false): string {
   const productInfo = text('Product Info', JSON.stringify(product));
 
   return codeSnippet === false