+++ packages/web-components/@sfx/sayt/src/sayt.ts
@@ -1,12 +1,17 @@
 import { LitElement, customElement, html, property, PropertyValues } from 'lit-element';
-import { PRODUCTS_RESPONSE_EVENT, PRODUCTS_REQUEST_EVENT } from '@sfx/products';
-import { SAYT_EVENT } from './events';
-import { SEARCHBOX_EVENT } from '@sfx/search-box';
 import {
-  AUTOCOMPLETE_RECEIVED_RESULTS_EVENT,
-  HOVER_AUTOCOMPLETE_TERM_EVENT,
-  AUTOCOMPLETE_REQUEST_RESULTS,
-} from '@sfx/autocomplete';
+  AUTOCOMPLETE_ACTIVE_TERM,
+  AUTOCOMPLETE_REQUEST,
+  AUTOCOMPLETE_RESPONSE,
+  SAYT_HIDE,
+  SAYT_SHOW,
+  SAYT_PRODUCTS_REQUEST,
+  SAYT_PRODUCTS_RESPONSE,
+  SEARCHBOX_INPUT,
+  AutocompleteActiveTermPayload,
+  SearchboxInputPayload,
+  WithGroup,
+} from '@sfx/events';
 
 /**
  * The `sfx-sayt` component is responsible for displaying and hiding the
@@ -69,11 +74,11 @@ export default class Sayt extends LitElement {
   connectedCallback() {
     super.connectedCallback();
 
-    window.addEventListener(SAYT_EVENT.SAYT_SHOW, this.showCorrectSayt);
-    window.addEventListener(AUTOCOMPLETE_RECEIVED_RESULTS_EVENT, this.showCorrectSayt);
-    window.addEventListener(PRODUCTS_RESPONSE_EVENT, this.showCorrectSayt);
-    window.addEventListener(SAYT_EVENT.SAYT_HIDE, this.hideCorrectSayt);
-    window.addEventListener(HOVER_AUTOCOMPLETE_TERM_EVENT, this.handleAutocompleteTermHover);
+    window.addEventListener(SAYT_SHOW, this.showCorrectSayt);
+    window.addEventListener(AUTOCOMPLETE_RESPONSE, this.showCorrectSayt);
+    window.addEventListener(SAYT_PRODUCTS_RESPONSE, this.showCorrectSayt);
+    window.addEventListener(SAYT_HIDE, this.hideCorrectSayt);
+    window.addEventListener(AUTOCOMPLETE_ACTIVE_TERM, this.handleAutocompleteTermHover);
     window.addEventListener('click', this.processClick);
     window.addEventListener('keydown', this.processKeyEvent);
     this.setSearchboxListener(this.searchbox, 'add');
@@ -85,11 +90,11 @@ export default class Sayt extends LitElement {
   disconnectedCallback() {
     super.disconnectedCallback();
 
-    window.removeEventListener(SAYT_EVENT.SAYT_SHOW, this.showCorrectSayt);
-    window.removeEventListener(AUTOCOMPLETE_RECEIVED_RESULTS_EVENT, this.showCorrectSayt);
-    window.removeEventListener(PRODUCTS_RESPONSE_EVENT, this.showCorrectSayt);
-    window.removeEventListener(SAYT_EVENT.SAYT_HIDE, this.hideCorrectSayt);
-    window.removeEventListener(HOVER_AUTOCOMPLETE_TERM_EVENT, this.handleAutocompleteTermHover);
+    window.removeEventListener(SAYT_SHOW, this.showCorrectSayt);
+    window.removeEventListener(AUTOCOMPLETE_RESPONSE, this.showCorrectSayt);
+    window.removeEventListener(SAYT_PRODUCTS_RESPONSE, this.showCorrectSayt);
+    window.removeEventListener(SAYT_HIDE, this.hideCorrectSayt);
+    window.removeEventListener(AUTOCOMPLETE_ACTIVE_TERM, this.handleAutocompleteTermHover);
     window.removeEventListener('click', this.processClick);
     window.removeEventListener('keydown', this.processKeyEvent);
     this.setSearchboxListener(this.searchbox, 'remove');
@@ -128,7 +133,7 @@ export default class Sayt extends LitElement {
       const searchbox = document.getElementById(searchboxId) as HTMLElement;
       if (searchbox) searchbox[setEventListener]('input', this.processSearchboxInput);
     } else {
-      window[setEventListener](SEARCHBOX_EVENT.SEARCHBOX_CHANGE, this.processSfxSearchboxChange);
+      window[setEventListener](SEARCHBOX_INPUT, this.processSfxSearchboxChange);
     }
   }
 
@@ -145,7 +150,7 @@ export default class Sayt extends LitElement {
    *
    * @param event An event that can contain a searchbox ID.
    */
-  showCorrectSayt(event: CustomEvent) {
+  showCorrectSayt(event: CustomEvent<WithGroup>) {
     if (this.isCorrectSayt(event)) {
       this.showSayt();
     }
@@ -163,7 +168,7 @@ export default class Sayt extends LitElement {
    *
    * @param event An event that can contain a searchbox ID.
    */
-  hideCorrectSayt(event: CustomEvent) {
+  hideCorrectSayt(event: CustomEvent<WithGroup>) {
     if (this.isCorrectSayt(event)) {
       this.hideSayt();
     }
@@ -193,34 +198,34 @@ export default class Sayt extends LitElement {
    *
    * @param eventType The type of the event to be dispatched.
    * @param query The query term.
-   * @param searchbox The ID of the associated searchbox.
+   * @param group The name of the associated group.
    */
-  dispatchRequestEvent(eventType: string, query: string, searchbox?: string) {
+  dispatchRequestEvent(eventType: string, query: string, group?: string) {
     const requestEvent = new CustomEvent(eventType, {
-      detail: { query, searchbox },
+      detail: { query, group },
       bubbles: true
     });
     this.dispatchEvent(requestEvent);
   }
 
   /**
-   * Dispatches an [[AUTOCOMPLETE_REQUEST_RESULTS]] event with the provided data.
+   * Dispatches an [[AUTOCOMPLETE_RESPONSE]] event with the provided data.
    *
    * @param query The search term to use.
    * @param searchbox The optional searchbox ID associated with this search.
    */
   requestSaytAutocompleteTerms(query: string, searchbox?: string) {
-    this.dispatchRequestEvent(AUTOCOMPLETE_REQUEST_RESULTS, query, searchbox);
+    this.dispatchRequestEvent(AUTOCOMPLETE_REQUEST, query, searchbox);
   }
 
   /**
-   * Dispatches a [[PRODUCTS_REQUEST_EVENT]] event with the provided data.
+   * Dispatches a [[SAYT_PRODUCTS_REQUEST]] event with the provided data.
    *
    * @param query The search term to use.
    * @param searchbox The optional searchbox ID associated with this search.
    */
   requestSaytProducts(query: string, searchbox?: string) {
-    this.dispatchRequestEvent(PRODUCTS_REQUEST_EVENT, query, searchbox);
+    this.dispatchRequestEvent(SAYT_PRODUCTS_REQUEST, query, searchbox);
   }
 
   /**
@@ -229,7 +234,7 @@ export default class Sayt extends LitElement {
    *
    * @param event The hover event dispatched from autocomplete.
    */
-  handleAutocompleteTermHover(event: CustomEvent) {
+  handleAutocompleteTermHover(event: CustomEvent<AutocompleteActiveTermPayload>) {
     this.requestSaytProducts(event.detail.query, this.searchbox);
   }
 
@@ -249,8 +254,8 @@ export default class Sayt extends LitElement {
    *
    * @param event The searchbox change event dispatched from the searchbox.
    */
-  processSfxSearchboxChange(event: CustomEvent) {
-    this.requestSayt(event.detail.value, event.detail.searchbox);
+  processSfxSearchboxChange(event: CustomEvent<SearchboxInputPayload>) {
+    this.requestSayt(event.detail.term, event.detail.group);
   }
 
   /**
@@ -260,9 +265,9 @@ export default class Sayt extends LitElement {
    *
    * @param event An event that can contain a searchbox ID for comparison.
    */
-  isCorrectSayt(event: CustomEvent): boolean {
-    const searchbox = event.detail && event.detail.searchbox;
-    return !searchbox || !this.searchbox || searchbox === this.searchbox;
+  isCorrectSayt(event: CustomEvent<WithGroup>): boolean {
+    const group = event.detail && event.detail.group;
+    return !group || !this.searchbox || group === this.searchbox;
   }
 
   /**