+++ webpack.config.js
@@ -25,7 +25,7 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /.(ts|tsx)?$/,
+        test: /\.tsx?$/,
         loader: 'ts-loader',
         exclude: [/node_modules/]
       },