+++ packages/web-components/@sfx/products/CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - SFX-200: Created the products wrapper.
   - This component renders a series of sfx-product components.
+  - SFX-248: The `@sfx/events` package is used for event names and payload interfaces.